pyproject.toml
[project]
name = "ai-policy-reader"
version = "1.0.0"
description = "AI 政策解读程序 — 基于大模型的多维度政策文件智能分析平台"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pdfplumber>=0.10.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]
